dicom: move dicom_json_conversion.py
